app/requirements.txt
flask==2.1.3
flask[async]
aioflask
neo4j
python-dotenv
loguru